Add files via upload
1)Combining Hotel Data from Different Years:
The code starts with a Common Table Expression (CTE) named Hotels which combines data from different years (2018, 2019, and 2020) of hotel records. Each year's data is stored in separate tables, and the UNION ALL operator is used to combine them.

2)Calculating Revenue and Total Days Stayed:
Within the Hotels CTE, additional columns are computed. The total days stayed (tot_days_stayed) is calculated by adding the number of weekend nights and week nights stayed. Revenue is computed by multiplying the total days stayed by the average daily rate (adr).

3)Analyzing Revenue Changes Yearwise:
The first SELECT statement outside the CTE is used to analyze the change in revenue for each hotel year-wise. It selects the arrival year, hotel name, and the sum of revenue for each hotel and year combination. The results are grouped by hotel and arrival year.

4)Joining with Market Segment and Meal Cost Tables:
The second part of the code uses another CTE named Hotels_with_MS_discount to combine hotel data with market segment and meal cost information.
It performs a LEFT JOIN operation with the market_segment$ table based on the market_segment column to incorporate market segment details into the hotel data.
Additionally, it performs a LEFT JOIN operation with the meal_cost$ table based on the meal column to include meal cost information in the hotel data.